==> notebooks/02_clean_transform.ipynb <==
#### ==============================================================
#### Instacart Project – Phase 2: Clean & Transform
#### - Load raw CSVs
#### - Concatenate prior + train order_products
#### - Filter out eval_set="test" from orders
#### - Save clean CSVs to data_clean
#### ==============================================================

import pandas as pd
import os

pd.set_option("display.max_columns", None)

#### --------------------------------------------------------------
#### Load raw data
#### --------------------------------------------------------------

DATA_PATH = r'data_raw'
DATA_CLEAN_PATH = r'data_clean'

os.listdir(DATA_PATH)

files = {
    "aisles": "aisles.csv",
    "departments": "departments.csv",
    "orders": "orders.csv",
    "order_products_prior": "order_products__prior.csv",
    "order_products_train": "order_products__train.csv",
    "products": "products.csv",
}

dfs = {}

for name, fname in files.items():
    path = os.path.join(DATA_PATH, fname)
    df = pd.read_csv(path)
    dfs[name] = df
    print(f"{name}: {df.shape[0]:,} rows × {df.shape[1]} columns")


## --------------------------------------------------------------
## Concatenate PRIOR + TRAIN into a single order_products table
## --------------------------------------------------------------

order_products_prior = dfs["order_products_prior"]
order_products_train = dfs["order_products_train"]

print("\norder_products_prior.info():")
print(order_products_prior.info())

order_products = pd.concat(
    [order_products_prior, order_products_train],
    axis=0,
    ignore_index=True
)

print("\nCombined order_products shape:", order_products.shape)

# Save combined order_products to data_clean
order_products.to_csv(
    os.path.join(DATA_CLEAN_PATH, "order_products.csv"),
    index=False
)


## --------------------------------------------------------------
## Clean orders table: remove eval_set = 'test'
## --------------------------------------------------------------

orders = dfs["orders"]

orders_rmv = orders[orders["eval_set"] != "test"]
print("\norders_rmv.info():")
print(orders_rmv.info())

orders_rmv.to_csv(
    os.path.join(DATA_CLEAN_PATH, "orders.csv"),
    index=False
)

## --------------------------------------------------------------
## quick join test in pandas (actual joins in PostgreSQL)
## --------------------------------------------------------------

orders_clean = orders_rmv.merge(order_products,on="order_id",how="left")

print("\norders_clean shape (for sanity check):", orders_clean.shape)
