==> notebooks/03_load_to_postgres.ipynb <==
# pyscr/load_to_postgres.py

Load cleaned Instacart CSVs into PostgreSQL tables.



import os
import pandas as pd
from sqlalchemy import create_engine



# -----------------------------
# 1. Database connection config
# -----------------------------
DB_USER = "postgres"
DB_PASS = "1998"
DB_HOST = "localhost"
DB_PORT = "5432"
DB_NAME = "instacart"
DB_SCHEMA = "instacart"

DB_URL = f"postgresql+psycopg2://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"

engine = create_engine(DB_URL)

# Base folder for cleaned CSVs
DATA_CLEAN_PATH = r'data_clean'




def load_csv(table_name: str, filename: str, dtype=None):
    """
    Helper to read a cleaned CSV and load it into a Postgres table.
    """
    path = os.path.join(DATA_CLEAN_PATH, filename)
    print(f"\nLoading {filename} into {DB_SCHEMA}.{table_name} ...")
    df = pd.read_csv(path, dtype=dtype)

    print(df.head())

    df.to_sql(
        name=table_name,
        con=engine,
        schema=DB_SCHEMA,
        if_exists="append",   
        index=False,
        chunksize=100_000,
        method="multi",
    )
    print(f"Loaded {len(df):,} rows into {DB_SCHEMA}.{table_name}")


def main():
    load_csv(
        table_name="departments",
        filename="departments.csv",
        dtype={
            "department_id": "int64",
            "department": "string",
        },
    )

    load_csv(
        table_name="aisles",
        filename="aisles.csv",
        dtype={
            "aisle_id": "int64",
            "aisle": "string",
        },
    )

    load_csv(
        table_name="products",
        filename="products.csv",
        dtype={
            "product_id": "int64",
            "product_name": "string",
            "aisle_id": "int64",
            "department_id": "int64",
        },
    )

    load_csv(
        table_name="orders",
        filename="orders.csv",
        dtype={
            "order_id": "int64",
            "user_id": "int64",
            "eval_set": "string",
            "order_number": "int64",
            "order_dow": "int64",
            "order_hour_of_day": "int64",
            "days_since_prior_order": "float64",
        },
    )


    load_csv(
        table_name="order_products",
        filename="order_products.csv",
        dtype={
            "order_id": "int64",
            "product_id": "int64",
            "add_to_cart_order": "int64",
            "reordered": "int64",   
        },
    )


if __name__ == "__main__":
    main()
